Karlson2k: Fix various warnings
git-svn-id: https://svn.code.sf.net/p/plibc/code/trunk/plibc@98 1596915d-9e13-0410-ab7f-96e59408e0f0

+++ src/mmap.c
@@ -244,12 +244,12 @@ int _win_munmap(void *start, size_t length)
 
     ReleaseMutex(hMappingsLock);
 
-    return success ? 0 : MAP_FAILED;
+    return success ? 0 : (int) MAP_FAILED;
   }
   else
   {
     SetErrnoFromWinError(GetLastError());
-    return MAP_FAILED;
+    return (int) MAP_FAILED;
   }
 }
